fix verify tests, and also move out forgot-password
