+++ auth-service/routes/auth-verify.test.ts
@@ -2,11 +2,15 @@ import { describe, it, expect, vi, beforeEach } from "vitest";
 import request from "supertest";
 import express from "express";
 import { createApp } from "../app.ts";
+import passport from "passport";
 
 describe("Verify Route", () => {
   let app: express.Express;
 
   beforeEach(() => {
+    (passport as any)._serializers = [];
+    (passport as any)._deserializers = [];
+
     app = createApp();
     vi.clearAllMocks();
   });
@@ -26,7 +30,9 @@ describe("Verify Route", () => {
     };
     const agent = request.agent(app);
     const res1 = await agent.post("/auth/register").send(userData);
+    expect(res1.status).toBe(200);
     const res2 = await agent.post("/auth/login").send(userData);
+    expect(res2.status).toBe(200);
     // Then verify authentication
     const response = await agent.get("/auth/verify");
     expect(response.status).toBe(200);
